status

+++ proj2/gitlet/Main.java
@@ -1,5 +1,7 @@
 package gitlet;
 
+import static gitlet.Repository.GITLET_DIR;
+
 /** Driver class for Gitlet, a subset of the Git version-control system.
  *  @author TODO
  */
@@ -10,42 +12,52 @@ public class Main {
      */
     public static void main(String[] args) {
         if (args.length < 1) {
-            throw new GitletException("Please enter a command.");
+            System.out.println("Please enter a command.");
+            return;
         }
         String firstArg = args[0];
+        if (!firstArg.equals("init") && !GITLET_DIR.exists()) {
+            System.out.println("Not in an initialized Gitlet directory.");
+            return;
+        }
         switch(firstArg) {
             case "init":
                 Repository.init();
                 break;
             case "add":
                 if (args.length != 2) {
-                    throw new GitletException("Usage: java gitlet.Main add [file name].");
+                    System.out.println("Usage: java gitlet.Main add [file name].");
+                    break;
                 }
                 Repository.add(args[1]);
                 break;
             case "rm":
                 if (args.length != 2) {
-                    throw new GitletException("Usage: java gitlet.Main remove [file name].");
+                    System.out.println("Usage: java gitlet.Main remove [file name].");
+                    break;
                 }
                 Repository.remove(args[1]);
                 break;
             case "commit":
                 if (args.length != 2) {
-                    throw new GitletException("Please enter a commit message.");
+                    System.out.println("Please enter a commit message.");
+                    break;
                 }
                 String message = args[1];
                 Repository.commit(message, "master");
                 break;
             case "branch":
                 if (args.length != 2) {
-                    throw new GitletException("Usage: java gitlet.Main branch [branch name].");
+                    System.out.println("Usage: java gitlet.Main branch [branch name].");
+                    break;
                 }
                 String branch = args[1];
                 Repository.branch(branch);
                 break;
             case "rm-branch":
                 if (args.length != 2) {
-                    throw new GitletException("Usage: java gitlet.Main rm-branch [branchname].");
+                    System.out.println("Usage: java gitlet.Main rm-branch [branchname].");
+                    break;
                 }
                 branch = args[1];
                 Repository.removeBranch(branch);
@@ -63,9 +75,18 @@ public class Main {
             case "log":
                 Repository.log();
                 break;
+            case "global-log":
+                Repository.globalLog();
+                break;
+            case "find":
+                String target = args[1];
+                Repository.find(target);
+                break;
+            case "status":
+                Repository.status();
+                break;
             default:
-                throw new GitletException("No command with that name exists.");
-            // TODO: FILL THE REST IN
+                System.out.println("No command with that name exists.");
         }
     }
 }
